Made it work
Login window now initializes first, and if login is correct it loads main window. Also removed a spelling error in the main window.
