--- src/presentation/LoginWindov.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane id="AnchorPane" prefHeight="200.0" prefWidth="350.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.141" fx:controller="presentation.LoginWindovController">
-   <children>
-      <PasswordField fx:id="passwordField" layoutX="101.0" layoutY="100.0" />
-      <TextField fx:id="usernameField" layoutX="101.0" layoutY="68.0" />
-      <Label layoutX="20.0" layoutY="68.0" text="Brugernavn" />
-      <Label layoutX="16.0" layoutY="104.0" text="Adgangskode" />
-      <Button fx:id="logInButton" layoutX="149.0" layoutY="146.0" mnemonicParsing="false" onAction="#loginButtonAction" text="Log ind" />
-   </children>
-</AnchorPane>
