--- src/presentation/LoginWindovController.java
@@ -1,54 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package presentation;
-
-import business.Runner;
-import java.io.IOException;
-import java.net.URL;
-import java.util.ResourceBundle;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Button;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
-
-/**
- * FXML Controller class
- *
- * @author nitra
- */
-public class LoginWindovController implements Initializable {
-    @FXML
-    private PasswordField passwordField;
-    @FXML
-    private TextField usernameField;
-    @FXML
-    private Button logInButton;
-    Runner r = new Runner();
-
-    /**
-     * Initializes the controller class.
-     */
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-        
-    }    
-
-    @FXML
-    private void loginButtonAction(ActionEvent event) {
-        if(r.login(usernameField.getText(), passwordField.getText()) == true) {
-            
-        } else {
-            Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setHeaderText("Fejl!");
-            alert.setContentText("Brugernavn eller password matcher ikke!");
-            alert.showAndWait();
-        }
-    }
-    
-}
