+++ src/presentation/LoginWindow.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane id="AnchorPane" fx:id="loginPane" prefHeight="140.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="presentation.LoginWindowController">
+   <children>
+      <TextField fx:id="usernameField" layoutX="99.0" layoutY="42.0" />
+      <PasswordField fx:id="passwordField" layoutX="99.0" layoutY="74.0" />
+      <Button fx:id="logInButton" layoutX="268.0" layoutY="55.0" mnemonicParsing="false" onAction="#loginButtonAction" onKeyPressed="#loginButtonKeyPressed" text="Log ind" />
+      <Label layoutX="18.0" layoutY="42.0" text="Brugernavn" />
+      <Label layoutX="14.0" layoutY="78.0" text="Adgangskode" />
+   </children>
+</AnchorPane>
