+++ src/presentation/UdredGUI.java
@@ -6,8 +6,6 @@ import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.TabPane;
-import javafx.scene.control.Tab;
 import javafx.stage.Stage;
 
 
@@ -16,11 +14,10 @@ import javafx.stage.Stage;
  * @author Bruger
  */
 public class UdredGUI extends Application {
-    private TabPane tabPane;
     
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("FXMLDocument.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("LoginWindow.fxml"));
         
         Scene scene = new Scene(root);
         
@@ -28,7 +25,7 @@ public class UdredGUI extends Application {
         stage.show();
         
     }
-
+    
     /**
      * @param args the command line arguments
      */
